Authenticate GitHub access in actions workflow (#95)
The default $GITHUB_TOKEN is useful only for this images repository;
it has no permissions for production-pipelines. Use a suitable PAT
for when the latter repo is, for example, temporarily made private
for operational reasons.

The secret name includes _RAW_ to indicate that its value must be
just the text of the PAT, not the "url= / username= / password="
format used in other repositories' workflows.

Unfortunately this line must be removed or commented out to revert
to using the default $GITHUB_TOKEN.

* Also use current GitHub Actions versions

While we're updating workflows, prevent "The following action(s) uses
node12 which is deprecated" warnings by updating to the current releases
of the Actions used.