Update dev environment to use ruff rather than black (#197)
* update dev environment to use ruff rather than black

to bring this repo in line with other software repos. We have switched
most things to ruff a while back to this upgrade is a little overdue. It
does result in formatting changes in some files but worth biting the
bullet and doing it now rather than waiting until we have more code in
here.