SET-483 Set up job to generate image statistics page on a schedule (#202)
Add script to run on a schedule to generate a page containing statistics about the usage of images.

This script pulls data from the artifact registry, and from a bigquery table that contains logs for pulls from the artifact registry. It then outputs that data as parquet files, and builds a static page using observable framework. This static page is then uploaded to the cpg-common main-web bucket where it can be served by the analysis runner web framework.

The intent of this page is to allow data managers to make informed decisions on the lifecycle of images, so they can see what images are in use by whom, and decide when certain images should be archived. 

---

Additionally there's some scaffolding work here to add a Makefile to simplify some installation and dependency compilation commands.