Add a new build process to new versioning convention (#191)
* trialing picard for new versioning

* trialing talos and samtools for image versioning

* working get version wfl script

* add checking archive repo

* fix linting

* added workflow to comment expected tags

* add content read perms to workflow

* removing unnecessary print statements

* checkout correct ref to extract changes

* add fromJson to correctly extract matrix

* remove fromJson and use direct reference

* store matrix in env

* replace manual comment with action

* added perms to create comments

* fix reading table from env

* use comment markdown template

* cleanup comment table text

* remove artifact from revamp workflow

* revert dockerfile changes

* revert talos Dockerfile

* revert to standard images repo

* disable on push event

* add on push trigger for revamp action

* Fix use of environment variable to fetch archive image tags

This correctly uses a different environment variable key for archived images instead of using the same for both prod and archived

* removing dummy image to test on push event

* change push event to feat branch for testing

* temporarily include images revamp branch in version compare

* adding dummy image to test images revamp versioning (#199)

* revert target to main branch post testing

* testing expected tags without gcloud auth

* testing expected tags with gcloud auth

* added verbosity for expected tags

* added verbosity in get version

* remove redundant verbosity from get version

* removing duplicated expected tags action

* fixed linting

* remove use of latest tag

* fix linting and bug in get_version

* Switch to images-tmp for testing

While we merge this PR, it might be safer to deploy to `images-tmp` on
the safer side, and we would revert it to use `images` once the tests
have concluded successfully.

* Fix bug in get_version

A bug in get_version caused an early return if either the prod or archive
registry returned no tags. This ensures both registries are checked and
the logic to get the next tag only runs after.

* Add logging the expected version to stderror

We do not have a way to debug any issues with the expected tag utility
as the json is not logged. This logs the json to stderror.

* Revert changes to cpg_workflows build process

The current process pushes a latest tag which we want to move away from
and the change would have removed that. This undoes that to keep it in
and address it separately.

---------

Co-authored-by: ckrishk <42167437+ckrishk@users.noreply.github.com>


* This PR adds a new build process for images

The changes here will ensure that the build process enforces the new versioning convention
and pushes tags in a deterministic way.