SET-465 Implement image management for data managers (#193)
* add logic to archive images based on comparing list to repository state

This builds on the work done by @violetbrina to "archive" images by
first copying them from the images repository to the images-archive
repository and then once successfully copied, deleting them from the
images repository.

The intent of this functionality is to ensure that out-of-date or
deprecated images are not accidentally used by researchers in production
analysis. By moving the images to the archive repository we can still
access them if necessary for provenance purposes but they cannot be
pulled in production workflows.

---------

Co-authored-by: Dan Coates <dan.coates@populationgenomics.org.au>