Add a new dev deploy workflow (#213)
* Add a new dev deploy workflow

The current workflow only supports pushing manually with workflow_dispatch
and also allows pushing to the prod registry which is unsafe. The new
workflow adds automatic pushes to the dev registry, and changes the
behaviour of the workflow_dispatch event to push to the dev registry only.

* Sanitise workflow_dispatch inputs

The workflow_dispatch call currently allows inputs that are not sanitised
and can probably be used to escape. This change aims to sanitise the input
so that the subsequent commands can be run safely.

* minor security fix

* Remove redundant step in deploy_images_auto

This step to build with args would never run due to conflicting
trigger comparisons with the job only running on a pull_request trigger
and the step only running on a workflow_dispatch trigger

* Use environment variables to process cli_args