Remove unnecessary CI step in revamped workflow (#216)
* Remove unnecessary CI step to build images for cpg_workflows

This basically copies in the changes from PR #184 to remove the step that
builds images for cpg_workflows as we no longer need it. Additional
details can be found on the PR above.

* Rename new workflows to make them distinct

Previous naming was very similar to existing workflows and made it
difficult to distinguish. New names contain "Revamped" for this purpose.