Label the image with a version number rather than branch name
The qtl fork doesn't tag releases, but there is a version number tracked
in DESCRIPTION. Use it, but add a (shortened) commit hash for clarity and
as it may not be reliably updated.

Remember to change the "Currently building from ..." comment when updating
this and returning to the main branch!