Update saige-qtl to 0.3.2.1-c2c9fb2 (#248)
* Update Dockerfile and label the image with a version number

The qtl fork doesn't tag releases, but there is a version number tracked
in DESCRIPTION. Use it, but add a (shortened) commit hash for clarity and
as it may not be reliably updated.

Remember to change the "Currently building from ..." comment when updating
this and returning to the main branch!

* Various patches so that the updated code compiles in our environment:

- Install savvy and superlu using driver-r's micromamba installation and
  pin libgcc-ng to avoid disrupting the existing installed packages.

- Replace pixi-based #! line with direct Rscript invocation

- Patch to add declaration of tbb::concurrent_vector
  It appears that RcppArmadillo 14.x included this header but 15.x does not.

---------

Co-authored-by: John Marshall <john.marshall@populationgenomics.org.au>